+++ prisma/schema.prisma
@@ -25,7 +25,7 @@ model Event {
   id                   String     @id @default(uuid())
   title                String
   date                 DateTime
-  image                String?
+  image                Bytes?
   style                String?
   description          String?
   reason               String?
