+++ src/actions/registerIser.ts
@@ -1,8 +1,8 @@
 "use server";
 
+import prisma from "@/../prisma/client";
 import bcrypt from "bcrypt";
 import { NextResponse } from "next/server";
-import prisma from "../../../prisma/client";
 
 export async function RegisterUser({
   email,
