+++ src/app/(pages)/events/[eventId]/page.tsx
