+++ src/app/(pages)/events/page.tsx
