+++ src/components/loginOrRegisterButtons/index.tsx
@@ -1,20 +1,17 @@
-import { AuthConfig } from "@/app/api/auth/[...nextauth]/config";
-import { getServerSession } from "next-auth";
+import getUserSession from "@/actions/getUserSession";
 import Link from "next/link";
 import { Button } from "../ui/button";
 import LoginButton from "./LoginButton";
+import LogOutButton from "./LogoutButton";
 
 const LoginOrRegisterButtons = async () => {
-  const session = await getServerSession(AuthConfig);
-
+  const session = await getUserSession();
   return (
     <div className="flex flex-col gap-4 row-start-1 items-center">
       {session ? (
         <div className="flex gap-4 justify-center items-center">
           <div className="font-bold">{session.user?.email}</div>
-          <Button variant="secondary">
-            <Link href="/api/auth/signout">Sign out</Link>
-          </Button>
+          <LogOutButton />
         </div>
       ) : (
         <div className="flex gap-4">
